Split _cog.py step 2
- Two step process to keep history
- rasterio and tiffile based implementation in separate files
- move some "converter" code into _fifffile as it's not generic enough
