+++ odc/geo/cog/__init__.py
@@ -1,7 +1,10 @@
 from ._rio import to_cog, write_cog, write_cog_layers
+from ._tifffile import CogMeta, save_cog_with_dask
 
 __all__ = [
+    "CogMeta",
     "to_cog",
     "write_cog",
     "write_cog_layers",
+    "save_cog_with_dask",
 ]
