+++ odc/geo/cog/_rio.py
@@ -5,27 +5,10 @@
 """
 Write Cloud Optimized GeoTIFFs from xarrays.
 """
-import itertools
 import warnings
 from contextlib import contextmanager
-from dataclasses import dataclass, field
-from functools import partial
-from io import BytesIO
 from pathlib import Path
-from time import monotonic
-from typing import (
-    Any,
-    Callable,
-    Dict,
-    Generator,
-    Iterable,
-    Iterator,
-    List,
-    Literal,
-    Optional,
-    Tuple,
-    Union,
-)
+from typing import Any, Dict, Generator, Iterable, List, Optional, Tuple, Union
 from uuid import uuid4
 
 import numpy as np
@@ -33,119 +16,12 @@ import rasterio
 import xarray as xr
 from rasterio.shutil import copy as rio_copy  # pylint: disable=no-name-in-module
 
-from ._interop import have
-from .converters import geotiff_metadata
-from .geobox import GeoBox
-from .math import align_down_pow2, align_up
-from .types import MaybeNodata, Shape2d, SomeShape, Unset, shape_, wh_
-from .warp import resampling_s2rio
+from ..geobox import GeoBox
+from ..math import align_up
+from ..types import MaybeNodata, SomeShape, shape_, wh_
+from ..warp import resampling_s2rio
 
 # pylint: disable=too-many-locals,too-many-branches,too-many-arguments,too-many-statements,too-many-instance-attributes
-# pylint: disable=too-many-lines
-
-AxisOrder = Union[Literal["YX"], Literal["YXS"], Literal["SYX"]]
-
-# map compressor name to level name in GDAL
-_GDAL_COMP: Dict[str, str] = {
-    "DEFLATE": "ZLEVEL",
-    "ADOBE_DEFLATE": "ZLEVEL",
-    "ZSTD": "ZSTD_LEVEL",
-    "WEBP": "WEBP_LEVEL",
-    "LERC": "MAX_Z_ERROR",
-    "LERC_DEFLATE": "MAX_Z_ERROR",
-    "LERC_ZSTD": "MAX_Z_ERROR",
-    "JPEG": "JPEG_QUALITY",
-}
-
-
-@dataclass
-class CogMeta:
-    """
-    COG metadata.
-    """
-
-    axis: AxisOrder
-    shape: Shape2d
-    tile: Shape2d
-    nsamples: int
-    dtype: Any
-    compression: int
-    predictor: int
-    compressionargs: Dict[str, Any] = field(default_factory=dict, repr=False)
-    gbox: Optional[GeoBox] = None
-    overviews: Tuple["CogMeta", ...] = field(default=(), repr=False)
-
-    def _pix_shape(self, shape: Shape2d) -> Tuple[int, ...]:
-        if self.axis == "YX":
-            return shape.shape
-        if self.axis == "YXS":
-            return (*shape.shape, self.nsamples)
-        return (self.nsamples, *shape.shape)
-
-    @property
-    def chunks(self) -> Tuple[int, ...]:
-        return self._pix_shape(self.tile)
-
-    @property
-    def pix_shape(self) -> Tuple[int, ...]:
-        return self._pix_shape(self.shape)
-
-    @property
-    def num_planes(self):
-        if self.axis == "SYX":
-            return self.nsamples
-        return 1
-
-    def flatten(self) -> Tuple["CogMeta", ...]:
-        return (self, *self.overviews)
-
-    @property
-    def chunked(self) -> Shape2d:
-        """
-        Shape in chunks.
-        """
-        ny, nx = ((N + n - 1) // n for N, n in zip(self.shape.yx, self.tile.yx))
-        return shape_((ny, nx))
-
-    @property
-    def num_tiles(self):
-        ny, nx = self.chunked.yx
-        return self.num_planes * ny * nx
-
-    def tidx(self) -> Iterator[Tuple[int, int, int]]:
-        """``[(plane_idx, iy, ix), ...]``"""
-        yield from np.ndindex((self.num_planes, *self.chunked.yx))
-
-    def flat_tile_idx(self, idx: Tuple[int, int, int]) -> int:
-        """Convert from sample,iy,ix to flat tile index."""
-        ns = self.num_planes
-        ny, nx = self.chunked.yx
-        for n, i in zip((ns, ny, nx), idx):
-            if i < 0 or i >= n:
-                raise IndexError()
-
-        sample, y, x = idx
-        return sample * (ny * nx) + y * nx + x
-
-    def cog_tidx(self) -> Iterator[Tuple[int, int, int, int]]:
-        """``[(ifd_idx, plane_idx, iy, ix), ...]``"""
-        idx_layers = list(enumerate(self.flatten()))[::-1]
-        for idx, mm in idx_layers:
-            yield from ((idx, pi, yi, xi) for pi, yi, xi in mm.tidx())
-
-    def __dask_tokenize__(self):
-        return (
-            "odc.CogMeta",
-            self.axis,
-            *self.shape.yx,
-            *self.tile.yx,
-            self.nsamples,
-            self.dtype,
-            self.compression,
-            self.predictor,
-            self.gbox,
-            len(self.overviews),
-        )
 
 
 def _without(cfg: Dict[str, Any], *skip: str) -> Dict[str, Any]:
@@ -199,50 +75,6 @@ def _num_overviews(block: int, dim: int) -> int:
         dim = dim // 2
         c += 1
     return c
-
-
-def _compute_cog_spec(
-    data_shape: SomeShape,
-    tile_shape: SomeShape,
-    *,
-    max_pad: Optional[int] = None,
-) -> Tuple[Shape2d, Shape2d, int]:
-    data_shape = shape_(data_shape)
-    tile_shape = shape_(shape_(tile_shape).map(_adjust_blocksize))
-    n1, n2 = (_num_overviews(b, dim) for dim, b in zip(data_shape.xy, tile_shape.xy))
-    n = max(n1, n2)
-    pad = 2**n
-    if max_pad is not None and max_pad < pad:
-        pad = 0 if max_pad == 0 else align_down_pow2(max_pad)
-
-    if pad > 0:
-        data_shape = shape_(data_shape.map(lambda d: align_up(d, pad)))
-    return (data_shape, tile_shape, n)
-
-
-def cog_gbox(
-    gbox: GeoBox,
-    *,
-    tile: Union[None, int, Tuple[int, int], Shape2d] = None,
-    nlevels: Optional[int] = None,
-) -> GeoBox:
-    """
-    Return padded gbox with safe dimensions for COG.
-
-    1. Compute number of desired overviews
-    2. Expand gebox on the right/bottom to have exact pixel shrink across all levels
-    """
-
-    if nlevels is None:
-        if tile is None:
-            tile = wh_(256, 256)
-        if isinstance(tile, int):
-            tile = wh_(tile, tile)
-        new_shape, _, _ = _compute_cog_spec(gbox.shape, tile)
-    else:
-        pad = 1 << nlevels
-        new_shape = shape_(gbox.shape.map(lambda d: align_up(d, pad)))
-    return gbox.expand(new_shape)
 
 
 def _default_cog_opts(
@@ -638,521 +470,3 @@ def write_cog_layers(
             else:
                 rio_copy(temp_fname, dst, copy_src_overviews=True, **rio_opts)
                 return Path(dst)
-
-
-def _yaxis_from_shape(
-    shape: Tuple[int, ...], gbox: Optional[GeoBox] = None
-) -> Tuple[AxisOrder, int]:
-    ndim = len(shape)
-
-    if ndim == 2:
-        return "YX", 0
-
-    if ndim != 3:
-        raise ValueError("Can only work with 2-d or 3-d data")
-    if shape[-1] in (3, 4):  # YXS in RGB(A)
-        return "YXS", 0
-
-    if gbox is None:
-        return "SYX", 1
-    if gbox.shape == shape[:2]:  # YXS
-        return "YXS", 0
-    if gbox.shape == shape[1:]:  # SYX
-        return "SYX", 1
-
-    raise ValueError("Geobox and image shape do not match")
-
-
-def _make_empty_cog(
-    shape: Tuple[int, ...],
-    dtype: Any,
-    gbox: Optional[GeoBox] = None,
-    *,
-    nodata: MaybeNodata = None,
-    gdal_metadata: Optional[str] = None,
-    compression: Union[str, Unset] = Unset(),
-    compressionargs: Any = None,
-    predictor: Union[int, bool, Unset] = Unset(),
-    blocksize: Union[int, List[Union[int, Tuple[int, int]]]] = 2048,
-    bigtiff: bool = True,
-    **kw,
-) -> Tuple[CogMeta, memoryview]:
-    # pylint: disable=import-outside-toplevel
-    have.check_or_error("tifffile", "rasterio", "xarray")
-    from tifffile import (
-        COMPRESSION,
-        FILETYPE,
-        PHOTOMETRIC,
-        PLANARCONFIG,
-        TiffWriter,
-        enumarg,
-    )
-
-    predictor, compression, compressionargs = _norm_compression_tifffile(
-        dtype,
-        predictor,
-        compression=compression,
-        compressionargs=compressionargs,
-        kw=kw,
-    )
-    _compression = enumarg(COMPRESSION, compression.upper())
-
-    if isinstance(blocksize, int):
-        blocksize = [blocksize]
-
-    ax, yaxis = _yaxis_from_shape(shape, gbox)
-    im_shape = shape_(shape[yaxis : yaxis + 2])
-    photometric = PHOTOMETRIC.MINISBLACK
-    planarconfig = PLANARCONFIG.SEPARATE
-    if ax == "YX":
-        nsamples = 1
-    elif ax == "YXS":
-        nsamples = shape[-1]
-        planarconfig = PLANARCONFIG.CONTIG
-        if nsamples in (3, 4):
-            photometric = PHOTOMETRIC.RGB
-    else:
-        nsamples = shape[0]
-
-    buf = BytesIO()
-
-    opts_common = {
-        "dtype": dtype,
-        "photometric": photometric,
-        "planarconfig": planarconfig,
-        "predictor": predictor,
-        "compression": _compression,
-        "compressionargs": compressionargs,
-        "software": False,
-        **kw,
-    }
-
-    def _sh(shape: Shape2d) -> Tuple[int, ...]:
-        if ax == "YX":
-            return shape.shape
-        if ax == "YXS":
-            return (*shape.shape, nsamples)
-        return (nsamples, *shape.shape)
-
-    tsz = _norm_blocksize(blocksize[-1])
-    im_shape, _, nlevels = _compute_cog_spec(im_shape, tsz)
-
-    extratags: List[Tuple[int, int, int, Any]] = []
-    if gbox is not None:
-        gbox = gbox.expand(im_shape)
-        extratags, _ = geotiff_metadata(
-            gbox, nodata=nodata, gdal_metadata=gdal_metadata
-        )
-    # TODO: support nodata/gdal_metadata without gbox?
-
-    _blocks = itertools.chain(iter(blocksize), itertools.repeat(blocksize[-1]))
-
-    tw = TiffWriter(buf, bigtiff=bigtiff, shaped=False)
-    metas: List[CogMeta] = []
-
-    for tsz, idx in zip(_blocks, range(nlevels + 1)):
-        tile = _norm_blocksize(tsz)
-        meta = CogMeta(
-            ax,
-            im_shape,
-            shape_(tile),
-            nsamples,
-            dtype,
-            int(_compression),
-            predictor,
-            compressionargs=compressionargs,
-            gbox=gbox,
-        )
-
-        if idx == 0:
-            kw = {**opts_common, "extratags": extratags}
-        else:
-            kw = {**opts_common, "subfiletype": FILETYPE.REDUCEDIMAGE}
-
-        tw.write(
-            itertools.repeat(b""),
-            shape=_sh(im_shape),
-            tile=tile,
-            **kw,
-        )
-
-        metas.append(meta)
-        im_shape = im_shape.shrink2()
-        if gbox is not None:
-            gbox = gbox.zoom_to(im_shape)
-
-    meta = metas[0]
-    meta.overviews = tuple(metas[1:])
-
-    tw.close()
-
-    return meta, buf.getbuffer()
-
-
-def _cog_block_compressor(
-    block: np.ndarray,
-    *,
-    tile_shape: Tuple[int, ...] = (),
-    encoder: Any = None,
-    predictor: Any = None,
-    axis: int = 1,
-    **kw,
-) -> bytes:
-    assert block.ndim == len(tile_shape)
-    if tile_shape != block.shape:
-        pad = tuple((0, want - have) for want, have in zip(tile_shape, block.shape))
-        block = np.pad(block, pad, "constant", constant_values=(0,))
-
-    if predictor is not None:
-        block = predictor(block, axis=axis)
-    if encoder:
-        try:
-            return encoder(block, **kw)
-        except Exception:  # pylint: disable=broad-except
-            return b""
-
-    return bytes(block.data)
-
-
-def mk_tile_compressor(meta: CogMeta) -> Callable[[np.ndarray], bytes]:
-    """
-    Make tile compressor.
-
-    """
-    # pylint: disable=import-outside-toplevel
-    have.check_or_error("tifffile")
-    from tifffile import TIFF
-
-    tile_shape = meta.chunks
-    encoder = TIFF.COMPRESSORS[meta.compression]
-
-    # TODO: handle SYX in planar mode
-    axis = 1
-    predictor = None
-    if meta.predictor != 1:
-        predictor = TIFF.PREDICTORS[meta.predictor]
-
-    return partial(
-        _cog_block_compressor,
-        tile_shape=tile_shape,
-        encoder=encoder,
-        predictor=predictor,
-        axis=axis,
-        **meta.compressionargs,
-    )
-
-
-def _compress_cog_tile(encoder, block, idx):
-    return [(idx, encoder(block))]
-
-
-def _compress_tiles(
-    xx: xr.DataArray,
-    meta: CogMeta,
-    scale_idx: int = 0,
-    sample_idx: int = 0,
-):
-    """
-    Compress chunks according to cog spec.
-
-    :returns: Dask bag of dicts with ``{"data": bytes, "idx": (int, int, int, int)}``
-    """
-    # pylint: disable=import-outside-toplevel
-    have.check_or_error("dask")
-    from dask.bag import Bag
-    from dask.base import quote, tokenize
-    from dask.highlevelgraph import HighLevelGraph
-
-    from ._interop import is_dask_collection
-
-    # TODO: deal with SYX planar data
-    assert meta.num_planes == 1
-    assert meta.axis in ("YX", "YXS")
-    assert meta.num_planes == 1
-    src_ydim = 0  # for now assume Y,X or Y,X,S
-
-    encoder = mk_tile_compressor(meta)
-    data = xx.data
-
-    if data.chunksize != meta.chunks:
-        data = data.rechunk(meta.chunks)
-
-    assert is_dask_collection(data)
-
-    tk = tokenize(
-        data,
-        scale_idx,
-        meta.axis,
-        meta.chunks,
-        meta.predictor,
-        meta.compression,
-        meta.compressionargs,
-    )
-    plane_id = "" if scale_idx == 0 else f"_{scale_idx}"
-    plane_id += "" if sample_idx == 0 else f"@{sample_idx}"
-
-    name = f"compress{plane_id}-{tk}"
-
-    src_data_name = data.name
-
-    def block_name(p, y, x):
-        if data.ndim == 2:
-            return (src_data_name, y, x)
-        if src_ydim == 0:
-            return (src_data_name, y, x, p)
-        return (src_data_name, p, y, x)
-
-    dsk = {}
-    for i, (p, y, x) in enumerate(meta.tidx()):
-        block = block_name(p, y, x)
-        dsk[name, i] = (_compress_cog_tile, encoder, block, quote((scale_idx, p, y, x)))
-
-    nparts = len(dsk)
-    dsk = HighLevelGraph.from_collections(name, dsk, dependencies=[data])
-    return Bag(dsk, name, nparts)
-
-
-def _pyramids_from_cog_metadata(
-    xx: xr.DataArray,
-    cog_meta: CogMeta,
-    resampling: Union[str, int] = "nearest",
-) -> Tuple[xr.DataArray, ...]:
-    out = [xx]
-
-    for mm in cog_meta.overviews:
-        gbox = mm.gbox
-        out.append(
-            out[-1].odc.reproject(gbox, chunks=mm.tile.yx, resampling=resampling)
-        )
-
-    return tuple(out)
-
-
-def _extact_tile_info(
-    meta: CogMeta,
-    tiles: List[Tuple[int, int, int, int, int]],
-    start_offset: int = 0,
-) -> List[Tuple[List[int], List[int]]]:
-    mm = meta.flatten()
-    tile_info = [([0] * m.num_tiles, [0] * m.num_tiles) for m in mm]
-
-    byte_offset = start_offset
-    for scale_idx, p, y, x, sz in tiles:
-        m = mm[scale_idx]
-        b_offsets, b_lengths = tile_info[scale_idx]
-
-        tidx = m.flat_tile_idx((p, y, x))
-        if sz != 0:
-            b_lengths[tidx] = sz
-            b_offsets[tidx] = byte_offset
-            byte_offset += sz
-
-    return tile_info
-
-
-def _update_header(
-    meta: CogMeta,
-    hdr0: bytes,
-    tiles: List[Tuple[int, int, int, int, int]],
-) -> bytes:
-    # pylint: disable=import-outside-toplevel
-    from tifffile import TiffFile
-
-    tile_info = _extact_tile_info(meta, tiles, len(hdr0))
-
-    _bio = BytesIO(hdr0)
-    with TiffFile(_bio, mode="r+", name=":mem:") as tr:
-        assert len(tile_info) == len(tr.pages)
-
-        # 324 -- offsets
-        # 325 -- byte counts
-        for info, page in zip(tile_info, tr.pages):
-            tags = page.tags
-            offsets, lengths = info
-            tags[324].overwrite(offsets)
-            tags[325].overwrite(lengths)
-
-    return bytes(_bio.getbuffer())
-
-
-def _norm_predictor(predictor: Union[int, bool, None], dtype: Any) -> int:
-    if predictor is False or predictor is None:
-        return 1
-
-    if predictor is True:
-        dtype = np.dtype(dtype)
-        if dtype.kind == "f":
-            return 3
-        if dtype.kind in "ui" and dtype.itemsize <= 4:
-            return 2
-        return 1
-    return predictor
-
-
-def _norm_compression_tifffile(
-    dtype: Any,
-    predictor: Union[bool, None, int, Unset] = Unset(),
-    compression: Union[str, Unset] = Unset(),
-    compressionargs: Any = None,
-    level: Optional[Union[int, float]] = None,
-    kw: Optional[Dict[str, Any]] = None,
-) -> Tuple[int, str, Dict[str, Any]]:
-    if kw is None:
-        kw = {}
-    if isinstance(compression, Unset):
-        compression = kw.pop("compress", "ADOBE_DEFLATE")
-        assert isinstance(compression, str)
-
-    if compressionargs is None:
-        compressionargs = {}
-
-    remap = {k.upper(): k for k in kw}
-
-    def opt(name: str, default=None) -> Any:
-        k = remap.get(name.upper(), None)
-        if k is None:
-            return default
-        return kw.pop(k, default)
-
-    def _gdal_level(compression: str, default=None) -> Any:
-        gdal_level_k = _GDAL_COMP.get(compression, None)
-        if gdal_level_k is None:
-            return default
-        return opt(gdal_level_k, default)
-
-    compression = compression.upper()
-
-    if level is None and "level" not in compressionargs:
-        # GDAL compat
-        level = _gdal_level(compression)
-
-    if level is not None:
-        compressionargs["level"] = level
-
-    if compression == "DEFLATE":
-        compression = "ADOBE_DEFLATE"
-    if compression == "LERC_DEFLATE":
-        compression = "LERC"
-        compressionargs["compression"] = "deflate"
-        if (lvl := _gdal_level("DEFLATE")) is not None:
-            compressionargs["compressionargs"] = {"level": lvl}
-    elif compression == "LERC_ZSTD":
-        compression = "LERC"
-        compressionargs["compression"] = "zstd"
-        if (lvl := _gdal_level("ZSTD")) is not None:
-            compressionargs["compressionargs"] = {"level": lvl}
-
-    if isinstance(predictor, Unset):
-        predictor = compression in ("ADOBE_DEFLATE", "ZSTD", "LZMA")
-
-    predictor = _norm_predictor(predictor, dtype)
-    return (predictor, compression, compressionargs)
-
-
-def save_cog_with_dask(
-    xx: xr.DataArray,
-    dst: str = "",
-    *,
-    compression: Union[str, Unset] = Unset(),
-    compressionargs: Any = None,
-    level: Optional[Union[int, float]] = None,
-    predictor: Union[int, bool, Unset] = Unset(),
-    blocksize: Union[Unset, int, List[Union[int, Tuple[int, int]]]] = Unset(),
-    bigtiff: bool = True,
-    overview_resampling: Union[int, str] = "nearest",
-    verbose: bool = False,
-    client: Any = None,
-    **kw,
-):
-    # pylint: disable=import-outside-toplevel
-    t0 = monotonic()
-    from dask import bag, delayed
-    from dask.utils import format_time
-
-    # usefull when debugging
-    optimize_graph = kw.pop("optimize_graph", True)
-
-    # normalize compression and remove GDAL compat options from kw
-    predictor, compression, compressionargs = _norm_compression_tifffile(
-        xx.dtype, predictor, compression, compressionargs, level=level, kw=kw
-    )
-    ydim = xx.odc.ydim
-    data_chunks: Tuple[int, int] = xx.data.chunksize[ydim : ydim + 2]
-    if isinstance(blocksize, Unset):
-        blocksize = [data_chunks, int(max(*data_chunks) // 2)]
-
-    meta, hdr0 = _make_empty_cog(
-        xx.shape,
-        xx.dtype,
-        xx.odc.geobox,
-        predictor=predictor,
-        compression=compression,
-        compressionargs=compressionargs,
-        blocksize=blocksize,
-        bigtiff=bigtiff,
-        nodata=xx.odc.nodata,
-        **kw,
-    )
-    hdr0 = bytes(hdr0)
-
-    layers = _pyramids_from_cog_metadata(xx, meta, resampling=overview_resampling)
-
-    _tiles = []
-    for scale_idx, (mm, img) in enumerate(zip(meta.flatten(), layers)):
-        _tiles.append(_compress_tiles(img, mm, scale_idx=scale_idx))
-
-    _tiles_merged = bag.concat(_tiles[::-1])
-    hdr_info = _tiles_merged.map(lambda d: (*d[0], len(d[1])))
-    tile_bytes = _tiles_merged.pluck(1)
-
-    new_hdr = delayed(_update_header, pure=True)(meta, hdr0, hdr_info)
-
-    dbg = {
-        "hdr_info": hdr_info,
-        "meta": meta,
-        "hdr0": hdr0,
-        "t0": t0,
-        "tiles_merged": _tiles_merged,
-    }
-
-    if dst == "":
-        return (new_hdr, tile_bytes, dbg)
-
-    if client is None:
-        return (new_hdr, tile_bytes, dbg)
-
-    def time_past() -> str:
-        return format_time(monotonic() - t0)
-
-    with open(dst, "wb") as fp:
-        if verbose:
-            print(f"[{time_past()}] Will write to: {dst}")
-            print(f"[{time_past()}] Starting computation on client: {client}]")
-
-        new_hdr, tile_bytes = client.persist(
-            [new_hdr, tile_bytes],
-            optimize_graph=optimize_graph,
-        )
-
-        if verbose:
-            print(f"[{time_past()}] Waiting for Dask to compress to RAM...")
-
-        new_hdr = client.compute(new_hdr, optimize_graph=optimize_graph)
-        new_hdr = new_hdr.result()  # blocks
-
-        if verbose:
-            print(f"[{time_past()}] DONE: compressed data is now in Dask")
-            print(f"... writing tiles to: {dst}")
-
-        fp.write(new_hdr)
-        for fut in tile_bytes.to_delayed():
-            for chunk in fut.compute():
-                fp.write(chunk)
-                if verbose:
-                    print(".", end="")
-        if verbose:
-            print(f"\n[{time_past()}] DONE")
-
-    dbg["t1"] = monotonic()
-    dbg["total_seconds"] = dbg["t1"] - dbg["t0"]
-    return new_hdr, tile_bytes, dbg
